Feature/UI settings (#14)
* refactor: storyboard 삭제, 코드 베이스로 변경

* chore: 주석 제거 및 네이밍 변경, 파일별 그룹화
