+++ BankManagerUIApp/BankManagerUIApp/Application/SceneDelegate.swift
@@ -1,0 +1,27 @@
+//
+//  BankManagerUIApp - SceneDelegate.swift
+//  Created by Eddy, dudu.
+//  Copyright © yagom academy. All rights reserved.
+// 
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        let mainViewController = BankViewController()
+        
+        window = UIWindow(windowScene: windowScene)
+        window?.rootViewController = mainViewController
+        window?.makeKeyAndVisible()
+    }
+
+    func sceneDidDisconnect(_ scene: UIScene) { }
+    func sceneDidBecomeActive(_ scene: UIScene) { }
+    func sceneWillResignActive(_ scene: UIScene) { }
+    func sceneWillEnterForeground(_ scene: UIScene) { }
+    func sceneDidEnterBackground(_ scene: UIScene) { }
+}
